test code optimization to use base hybrid test file
