+++ autorag/nodes/retrieval/hybrid_rsf.py
@@ -12,6 +12,8 @@ def hybrid_rsf(ids: Tuple, scores: Tuple, top_k: int,
                weights: Tuple = (0.5, 0.5)) -> Tuple[List[str], List[float]]:
     assert len(ids) == len(scores), "The length of ids and scores must be the same."
     assert len(ids) > 1, "You must input more than one retrieval results."
+    assert top_k > 0, "top_k must be greater than 0."
+    assert sum(weights) == 1, "The sum of weights must be 1."
 
     # Initialize DataFrame for ids and scores
     id_df = pd.DataFrame({f'id_{i}': id_list for i, id_list in enumerate(ids)})
@@ -19,12 +21,12 @@ def hybrid_rsf(ids: Tuple, scores: Tuple, top_k: int,
     df = pd.concat([id_df, score_df], axis=1)
 
     # Apply relative score fusion
-    def rsf_apply(row):
+    def rsf_pure_apply(row):
         ids_tuple = tuple(row[[f'id_{i}' for i in range(len(ids))]].values)
         scores_tuple = tuple(row[[f'score_{i}' for i in range(len(scores))]].values)
         return pd.Series(rsf_pure(ids_tuple, scores_tuple, top_k, dist_based, weights))
 
-    df[['rsf_id', 'rsf_score']] = df.apply(rsf_apply, axis=1)
+    df[['rsf_id', 'rsf_score']] = df.apply(rsf_pure_apply, axis=1)
     return df['rsf_id'].tolist(), df['rsf_score'].tolist()
